# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Creates a new user account and returns JWT token in httpOnly cookie",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered"
          },
          "400": {
            "description": "Validation error or email already exists"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates user and returns JWT token in httpOnly cookie",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Returns the currently authenticated user information",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user data"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Clears the JWT token cookie",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          }
        },
        "summary": "Logout user",
        "tags": [
          "auth"
        ]
      }
    },
    "/quiz/{id}/start": {
      "post": {
        "description": "Creates and starts a new quiz session (Teacher only)",
        "operationId": "QuizController_startSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quiz ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Session created successfully"
          },
          "403": {
            "description": "Only teachers can start sessions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start quiz session",
        "tags": [
          "quizzes"
        ]
      }
    },
    "/quiz": {
      "post": {
        "description": "Creates a new quiz with questions (Teacher only)",
        "operationId": "QuizController_createQuiz",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quiz created successfully"
          },
          "403": {
            "description": "Only teachers can create quizzes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create quiz",
        "tags": [
          "quizzes"
        ]
      },
      "get": {
        "operationId": "QuizController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/quiz/{id}": {
      "get": {
        "description": "Retrieves a quiz with all questions and answers",
        "operationId": "QuizController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quiz ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz found"
          },
          "404": {
            "description": "Quiz not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get quiz by ID",
        "tags": [
          "quizzes"
        ]
      },
      "put": {
        "operationId": "QuizController_updateQuiz",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      },
      "delete": {
        "operationId": "QuizController_deleteQuiz",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quizzes"
        ]
      }
    },
    "/session/join": {
      "post": {
        "description": "Join an active quiz session using a session code",
        "operationId": "SessionController_join",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully joined session"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join quiz session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}": {
      "get": {
        "operationId": "SessionController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}/submit": {
      "post": {
        "operationId": "SessionController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}/advance": {
      "post": {
        "operationId": "SessionController_advance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}/state": {
      "get": {
        "operationId": "SessionController_state",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}/start": {
      "post": {
        "operationId": "SessionController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}/quiz": {
      "get": {
        "operationId": "SessionController_quiz",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}/results": {
      "get": {
        "operationId": "SessionController_results",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}/finish": {
      "post": {
        "operationId": "SessionController_finish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    },
    "/session/{id}/question-stats": {
      "get": {
        "operationId": "SessionController_questionStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "sessions"
        ]
      }
    }
  },
  "info": {
    "title": "Real-Time Quiz API",
    "description": "API for interactive real-time quiz platform with WebSocket support",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "auth",
      "description": "Authentication endpoints"
    },
    {
      "name": "users",
      "description": "User management"
    },
    {
      "name": "quizzes",
      "description": "Quiz CRUD operations"
    },
    {
      "name": "questions",
      "description": "Question management"
    },
    {
      "name": "sessions",
      "description": "Quiz session management"
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "lastName": {
            "type": "string",
            "example": "Dupont",
            "description": "User last name"
          },
          "firstName": {
            "type": "string",
            "example": "Jean",
            "description": "User first name"
          },
          "email": {
            "type": "string",
            "example": "jean.dupont@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password (min 6 characters)",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "TEACHER",
              "STUDENT"
            ],
            "example": "STUDENT",
            "description": "User role (TEACHER or STUDENT)"
          }
        },
        "required": [
          "lastName",
          "firstName",
          "email",
          "password",
          "role"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jean.dupont@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateQuizRequest": {
        "type": "object",
        "properties": {}
      },
      "JoinSessionRequest": {
        "type": "object",
        "properties": {}
      },
      "SubmitAnswerRequest": {
        "type": "object",
        "properties": {}
      },
      "AdvanceSessionRequest": {
        "type": "object",
        "properties": {}
      }
    }
  }
}